requirements.txt
boto3==1.12.18
mysql-connector==2.2.9